Update patch set 2
Patch Set 2:

> Patch Set 2: Verified-1 Code-Review-1
> 
> Cranks up the frequencies even when not being used much.
> Plus, increased the battery drain

Strange. Well I suggested this because your jitter was in the 1ms range with post_boot script so I suspected that might be because your cpu was not ramping up when required. It happened in my device when use_sched_load was enabled (introduced lags and jitter was more than 2ms)

Patch-set: 2